backprop param, debug-compatible, zero out all weights for now, turn off PM task
right now it learns the FFWD task only, without any attention, without a
PM task
also prints nice figures with the activation right before the last trial
of learning (it also stops when the marginal hit rate is 90%)

no LAMBDA (normalization) yet
the weights are not bad though

one weird thing is that it learns to always suppress one output and
default to the other (e.g. the biases for outputs are -2 and 2) and then
switches to the non-default one. Not sure how to add symmetry yet. Some
ideas -- weight = average weight for group of connections, etc.